+++ Harden-Windows-Security Module/Main files/Resources/XAML/ResourceDictionaries/Text.xaml
@@ -1,0 +1,22 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+    <Style x:Key="Text_Style"
+           TargetType="TextBlock">
+
+        <Setter Property="FontFamily"
+                Value="/Fonts/#Rubik" />
+        <Setter Property="VerticalAlignment"
+                Value="Center" />
+        <Setter Property="HorizontalAlignment"
+                Value="Left" />
+        <Setter Property="Foreground"
+                Value="#DBDBDB" />
+        <Setter Property="FontSize"
+                Value="18" />
+        <Setter Property="Margin"
+                Value="70,0,0,0" />
+
+    </Style>
+
+</ResourceDictionary>
